switch to next consumer upon receiving a nil message
The EOF Partition error is only received when a topic is read to the end,
but only if there have been new messages in the meantime. Put differently,
if a topic receives no messages no EOF will be raised apart from the
initial one.

If a topic thus receives no messages whatsoever, ConsumerSet would wait
indefinitely for this consumer, even if there are messages on other
subscribed topics. This patch changes the behaviour to switch to the next
consumer upon receiving a nil message.